#1294 Cannot set analysis log size to Off  Return 0 when we fail to parse
 #931 Missing completions for express
 Improve modeling of this across calls
 Also limits linked value propagation to 5 types to prevent this from exploding the analysis

 #1097 Goto def on field assign doesn't go to definition
 Report assignments as definitions when getting variables

 Also:
 Use TypesNoCopy instead of GetTypesNoCopy to avoid visibility filtering when getting definitions
 Don't delete analysis on implicit project (which crashes VS)

+++ Nodejs/Product/Nodejs/Intellisense/VsProjectAnalyzer.cs
@@ -1256,11 +1256,13 @@ namespace Microsoft.NodejsTools.Intellisense {
         }
 
         private void DeleteAnalysis() {
-            try {
-                var path = GetAnalysisPath();
-                File.Delete(path);
-            } catch (IOException) {
-            } catch (UnauthorizedAccessException) {
+            if (_projectDir != null) {
+                try {
+                    var path = GetAnalysisPath();
+                    File.Delete(path);
+                } catch (IOException) {
+                } catch (UnauthorizedAccessException) {
+                }
             }
         }
